Add different recipes for this nemo 2.0 example
